refactor(operations): make the editor's tabs sibling paths, not a ?tab= query
`/operations/{id}` (public orders) · `/edit` (Brief) · `/map` · `/schedule` ·
`/attendance`. They are four views of one operation, so they are siblings; the
first attempt nested them under `/edit`, which read as if Brief were the
operation and the rest were sub-pages of the editor.

**Switching tabs still never navigates.** `useEditorTab` already mirrored the
tab into the URL with `replaceState` rather than `router.push`, because a real
navigation tears down the Hocuspocus socket and forces the Y.Doc to reconnect on
every switch. That is unchanged — only the shape of the URL it writes. The route
files exist purely to answer a cold load or a pasted link, which is why the URL
can cross what Next considers different routes without the socket noticing.

`?tab=` is still read, since it is in bookmarks and Discord messages, and is then
normalised away into the path form. Leaving both would mean the next switch left
a query contradicting the path, and a copied link opening a different tab than
the one on screen.

`/operations/{id}/map` was already taken by the public fullscreen viewer, so that
path now serves whichever version the viewer is entitled to — the editor's Map
tab for anyone who can edit, the read-only viewer for everyone else. Splitting it
into two paths would mean the link people paste each other only works for half of
them.

`/edit` has a layout that gates it; the new siblings sit outside that layout and
carry their own check. A member without edit rights lands on the operation's
public page rather than the operations list — they asked for *this* operation.

Six E2E specs added: cold loads of each path, the switch rewriting the URL, the
legacy query being normalised, and the redirect for a member. Not run — the
suite is the user's to run.
